More than words

+++ 0x1E-search_algorithms/100-jump.c
@@ -35,7 +35,7 @@ int jump_search(int *array, size_t size, int value)
 	printf("Value checked array[%lu] = [%d]\n", a, array[a]);
 
 	if (array[a] == value)
-		return (a);
+		return ((int)a);
 
 	return (-1);
 }
